pyproject.toml
[project]
name = "tucosto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.42.1",
    "gspread>=6.2.1",
    "polars[excel]>=1.35.1",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.4",
    "streamlit>=1.51.0",
]
